requirements.txt
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.0.76
albumentations==1.3.1
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
scikit-learn==1.3.0
Pillow==10.0.0
streamlit==1.28.0
kaggle==1.5.16
anthropic==0.20.0
tqdm==4.66.1
scipy==1.11.3
